🐛 FIX(notifications): General small refactors
