+++ lib/pages/catalog/product/info/blocks/product_info_card/product_price_card.dart
@@ -1,14 +1,12 @@
 // Flutter imports:
-import 'package:flutter/material.dart';
-
 // Package imports:
 import 'package:calendar_date_picker2/calendar_date_picker2.dart';
-
 // Project imports:
 import 'package:exdock_backend_client/globals/styling.dart';
-import 'package:exdock_backend_client/pages/catalog/attributes/generate_attribute.dart';
 import 'package:exdock_backend_client/pages/catalog/product/info/product_info_card/product_info_card_title.dart';
+import 'package:exdock_backend_client/utils/attributes/generate_attribute.dart';
 import 'package:exdock_backend_client/utils/map_notifier.dart';
+import 'package:flutter/material.dart';
 
 class ProductPriceCard extends StatefulWidget {
   const ProductPriceCard({
