+++ lib/pages/catalog/product/info/product_info.dart
@@ -1,16 +1,13 @@
 // Dart imports:
 import 'dart:convert';
 
-// Flutter imports:
-import 'package:flutter/material.dart';
-
 // Project imports:
 import 'package:exdock_backend_client/pages/catalog/product/info/product_info_synchronous.dart';
-import 'package:exdock_backend_client/pages/catalog/product/info/product_info_synchronous.dart';
-import 'package:exdock_backend_client/utils/HTTP/HttpData.dart';
+import 'package:exdock_backend_client/utils/HTTP/http_data.dart';
 import 'package:exdock_backend_client/utils/HTTP/post_requests.dart';
-import 'package:exdock_backend_client/utils/MapNotifier.dart';
 import 'package:exdock_backend_client/utils/map_notifier.dart';
+// Flutter imports:
+import 'package:flutter/material.dart';
 
 class ProductInfo extends StatefulWidget {
   const ProductInfo({
