+++ lib/utils/HTTP/connect_websocket_stream.dart
@@ -1,23 +1,21 @@
 // Dart imports:
 import 'dart:convert';
 
+// Project imports:
+import 'package:exdock_backend_client/utils/HTTP/login_requests.dart';
+import 'package:exdock_backend_client/utils/authentication/authentication_data.dart';
 // Flutter imports:
 import 'package:flutter/cupertino.dart';
-
 // Package imports:
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:web_socket_channel/web_socket_channel.dart';
 
-// Project imports:
-import 'package:exdock_backend_client/utils/HTTP/login_requests.dart';
-import 'package:exdock_backend_client/utils/authentication/authentication_data.dart';
-
 void getWebsocketChannel(String url, ValueNotifier values) async {
-  FlutterSecureStorage storage = FlutterSecureStorage();
+  FlutterSecureStorage storage = const FlutterSecureStorage();
   String? accessToken = await storage.read(key: "access_token");
   String? refreshToken = await storage.read(key: "refresh_token");
-  final wsUrl = Uri.parse(url);
-  final channel = WebSocketChannel.connect(wsUrl);
+  var wsUrl = Uri.parse(url);
+  var channel = WebSocketChannel.connect(wsUrl);
   bool isFirstAttempt = true;
   bool isAuthenticated = false;
