+++ lib/utils/HTTP/login_requests.dart
@@ -1,19 +1,17 @@
 // Dart imports:
 import 'dart:convert';
 
+// Project imports:
+import 'package:exdock_backend_client/utils/HTTP/http_data.dart';
+import 'package:exdock_backend_client/utils/authentication/authentication_data.dart';
 // Flutter imports:
 import 'package:flutter/foundation.dart';
-
 // Package imports:
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:http/http.dart' as http;
 
-// Project imports:
-import 'package:exdock_backend_client/utils/HTTP/HttpData.dart';
-import 'package:exdock_backend_client/utils/authentication/authentication_data.dart';
-
 Future<HttpData> loginRequest(AuthenticationData authData) async {
-  FlutterSecureStorage storage = FlutterSecureStorage();
+  FlutterSecureStorage storage = const FlutterSecureStorage();
   int statusCode = 404;
   String baseUrl = "http://127.0.0.1:80";
 
@@ -49,7 +47,7 @@ Future<HttpData> loginRequest(AuthenticationData authData) async {
 }
 
 Future<String> refreshTokens(String refreshToken) async {
-  FlutterSecureStorage storage = FlutterSecureStorage();
+  FlutterSecureStorage storage = const FlutterSecureStorage();
   // TODO: add server from config
   String baseUrl = "http://127.0.0.1";
   Uri uri = Uri.parse("$baseUrl/api/v1/refresh");
