+++ lib/utils/HTTP/post_requests.dart
@@ -1,18 +1,17 @@
 // Dart imports:
 import 'dart:io';
 
+// Project imports:
+import 'package:exdock_backend_client/utils/HTTP/http_data.dart';
+import 'package:exdock_backend_client/utils/HTTP/login_requests.dart';
+import 'package:exdock_backend_client/utils/authentication/authentication_data.dart';
 // Package imports:
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:http/http.dart' as http;
 import 'package:jwt_decoder/jwt_decoder.dart';
 
-// Project imports:
-import 'package:exdock_backend_client/utils/HTTP/HttpData.dart';
-import 'package:exdock_backend_client/utils/authentication/authentication_data.dart';
-import 'login_requests.dart';
-
 Future<HttpData> standardPostRequest(String endpoint, String body) async {
-  FlutterSecureStorage storage = FlutterSecureStorage();
+  FlutterSecureStorage storage = const FlutterSecureStorage();
   int statusCode = 404;
   String responseBody = "";
   String? accessToken = await storage.read(key: "access_token");
