+++ lib/utils/attributes/generate_attribute.dart
@@ -1,13 +1,12 @@
 // Flutter imports:
-import 'package:flutter/material.dart';
-
 // Project imports:
-import 'package:exdock_backend_client/pages/catalog/attributes/images/single_image_attribute.dart';
-import 'package:exdock_backend_client/pages/catalog/attributes/text_field_attribute.dart';
-import 'package:exdock_backend_client/pages/catalog/attributes/text_field_number_attribute.dart';
-import 'package:exdock_backend_client/pages/catalog/attributes/text_field_price_attribute.dart';
-import 'package:exdock_backend_client/pages/catalog/attributes/wysiwyg_attribute.dart';
+import 'package:exdock_backend_client/utils/attributes/images/single_image_attribute.dart';
+import 'package:exdock_backend_client/utils/attributes/text_field_attribute.dart';
+import 'package:exdock_backend_client/utils/attributes/text_field_number_attribute.dart';
+import 'package:exdock_backend_client/utils/attributes/text_field_price_attribute.dart';
+import 'package:exdock_backend_client/utils/attributes/wysiwyg_attribute.dart';
 import 'package:exdock_backend_client/utils/map_notifier.dart';
+import 'package:flutter/material.dart';
 
 class GenerateAttribute extends StatelessWidget {
   const GenerateAttribute(
