+++ lib/utils/attributes/images/single_image_attribute.dart
@@ -1,13 +1,11 @@
 // Flutter imports:
-import 'package:flutter/material.dart';
-
 // Package imports:
 import 'package:dotted_border/dotted_border.dart';
-import 'package:kumi_popup_window/kumi_popup_window.dart';
-
 // Project imports:
-import 'package:exdock_backend_client/pages/catalog/attributes/images/single_image_attribute_popup.dart';
+import 'package:exdock_backend_client/utils/attributes/images/single_image_attribute_popup.dart';
 import 'package:exdock_backend_client/utils/map_notifier.dart';
+import 'package:flutter/material.dart';
+import 'package:kumi_popup_window/kumi_popup_window.dart';
 
 class SingleImageAttribute extends StatefulWidget {
   const SingleImageAttribute(
