+++ lib/utils/blocks/group_cards/category_edit_group_card.dart
@@ -1,9 +1,8 @@
 // Flutter imports:
-import 'package:flutter/material.dart';
-
 // Project imports:
 import 'package:exdock_backend_client/globals/globals.dart';
-import 'package:exdock_backend_client/pages/catalog/category/edit/content/group_cards/unsaved_changes_overlay.dart';
+import 'package:exdock_backend_client/pages/catalog/product/info/product_info_card/unsaved_changes_overlay.dart';
+import 'package:flutter/material.dart';
 
 class CategoryEditGroupCard extends StatelessWidget {
   const CategoryEditGroupCard({
