+++ lib/utils/blocks/standard_block.dart
@@ -1,10 +1,9 @@
 // Flutter imports:
-import 'package:flutter/material.dart';
-
 // Project imports:
-import 'package:exdock_backend_client/pages/catalog/attributes/generate_attribute.dart';
-import 'package:exdock_backend_client/pages/catalog/category/edit/content/group_cards/category_edit_group_card_with_title.dart';
+import 'package:exdock_backend_client/utils/attributes/generate_attribute.dart';
+import 'package:exdock_backend_client/utils/blocks/group_cards/category_edit_group_card_with_title.dart';
 import 'package:exdock_backend_client/utils/map_notifier.dart';
+import 'package:flutter/material.dart';
 
 class StandardBlock extends StatefulWidget {
   const StandardBlock(
